+++ src/egon/data/datasets/power_plants/__init__.py
@@ -17,7 +17,7 @@ import geopandas as gpd
 import numpy as np
 import pandas as pd
 
-from egon.data import db
+from egon.data import db, config
 from egon.data.datasets import Dataset
 from egon.data.datasets.power_plants.pv_rooftop import pv_rooftop_per_mv_grid
 from egon.data.datasets.power_plants.conventional import (
@@ -74,7 +74,7 @@ def create_tables():
     None.
     """
 
-    cfg = egon.data.config.datasets()["power_plants"]
+    cfg = config.datasets()["power_plants"]
     db.execute_sql(f"CREATE SCHEMA IF NOT EXISTS {cfg['target']['schema']};")
     engine = db.engine()
     db.execute_sql(
@@ -137,7 +137,7 @@ def select_target(carrier, scenario):
         Target values for carrier and scenario
 
     """
-    cfg = egon.data.config.datasets()["power_plants"]
+    cfg = config.datasets()["power_plants"]
 
     return (
         pd.read_sql(
@@ -175,7 +175,7 @@ def filter_mastr_geometry(mastr, federal_state=None):
         Power plants listed in MaStR with geometry inside German boundaries
 
     """
-    cfg = egon.data.config.datasets()["power_plants"]
+    cfg = config.datasets()["power_plants"]
 
     if type(mastr) == pd.core.frame.DataFrame:
         # Drop entries without geometry for insert
@@ -231,7 +231,7 @@ def insert_biomass_plants(scenario):
     None.
 
     """
-    cfg = egon.data.config.datasets()["power_plants"]
+    cfg = config.datasets()["power_plants"]
 
     # import target values from NEP 2021, scneario C 2035
     target = select_target("biomass", scenario)
@@ -313,7 +313,7 @@ def insert_hydro_plants(scenario):
     None.
 
     """
-    cfg = egon.data.config.datasets()["power_plants"]
+    cfg = config.datasets()["power_plants"]
 
     # Map MaStR carriers to eGon carriers
     map_carrier = {
@@ -574,7 +574,7 @@ def insert_hydro_biomass():
     None.
 
     """
-    cfg = egon.data.config.datasets()["power_plants"]
+    cfg = config.datasets()["power_plants"]
     db.execute_sql(
         f"""
         DELETE FROM {cfg['target']['schema']}.{cfg['target']['table']}
@@ -591,7 +591,7 @@ def allocate_conventional_non_chp_power_plants():
 
     carrier = ["oil", "gas", "other_non_renewable"]
 
-    cfg = egon.data.config.datasets()["power_plants"]
+    cfg = config.datasets()["power_plants"]
 
     for carrier in carrier: