+++ CHANGELOG.rst
@@ -139,6 +139,8 @@ Added
   `#359 <https://github.com/openego/eGon-data/issues/359>`_
 * Include allocation of conventional (non CHP) power plants
   `#392 <https://github.com/openego/eGon-data/issues/392>`_
+* Include allocation of pumped hydro units
+  `#332 <https://github.com/openego/eGon-data/issues/332>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159